requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1

django-filter~=24.2
python-dotenv~=1.0.1
stripe~=10.2.0
requests~=2.32.3
celery~=5.4.0